+++ src/main/java/com/example/r2dbmsreactiveschool/repositories/StudentRepository.java
@@ -1,11 +1,11 @@
 package com.example.r2dbmsreactiveschool.repositories;
 
-import com.example.r2dbmsreactiveschool.domian.Student;
+import com.example.r2dbmsreactiveschool.domain.Student;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentRepository extends ReactiveCrudRepository<Student, Integer> {
+public interface StudentRepository extends ReactiveCrudRepository<Student, Long> {
 
 
 }
